add dateDivider, fix buid

+++ androidApp/src/main/java/com/socialsirius/messenger/ui/chats/chat/item/IChatItem.kt
@@ -10,7 +10,7 @@ open class IChatItem() {
         return ""
     }
 
-   open var timeInMillis: Double =0.0
+   open var timeInMillis: Long = 0
 
     companion object {
 /*
