+++ androidApp/src/main/java/com/socialsirius/messenger/ui/chats/chat/item/TechMessageItem.kt
@@ -6,7 +6,7 @@ class TechMessageItem private constructor(
     val text: String,
     val dateTime: String,
     val id: String,
-    override var timeInMillis: Double
+    override var timeInMillis: Long
 ) : IChatItem() {
     companion object {
   /*      @JvmStatic
